db-models/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "db-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "db-models",
    "psycopg[binary]>=3.2.12",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
]

[tool.uv.sources]
db-models = { workspace = true }
